#1356: split bpf_map.rs publish_bpf_conntrack_entry per-AF
Pure code motion. The 204-LOC publish_bpf_conntrack_entry function in
userspace-dp/src/afxdp/bpf_map.rs branched on key.addr_family and inlined
two parallel v4/v6 BPF key/value construction arms (lines 449-651 on
master). Both arms moved to private per-family helpers under a new
afxdp/bpf_map/publish_conntrack.rs submodule. The orchestrator stays in
bpf_map/mod.rs, dispatches by address family, and now sits at 56 LOC.

docs/engineering-style.md defines the soft cap at 100 LOC per function
and the hard Tier-1 cap at 200 LOC. The original body was >4x the soft
cap and >2x the hard cap; the new orchestrator + each per-family helper
all sit well under the soft cap.

Mechanical changes:

- git mv userspace-dp/src/afxdp/bpf_map.rs
        userspace-dp/src/afxdp/bpf_map/mod.rs

- userspace-dp/src/afxdp/mod.rs:59-60: update path attribute from
  #[path = "bpf_map.rs"] to #[path = "bpf_map/mod.rs"] per the
  project's explicit-path convention (matches forwarding/mod.rs,
  frame/mod.rs, etc.).

- bpf_map/mod.rs tail: update the test include from
  #[path = "bpf_map_tests.rs"] to #[path = "../bpf_map_tests.rs"]
  so the existing sibling test file resolves after the directory move.

- bpf_map/mod.rs body: shrink publish_bpf_conntrack_entry from 204 LOC
  to a 56-LOC orchestrator that computes zone IDs, now_secs, and flags
  once, then dispatches to publish_conntrack::{publish_v4_session,
  publish_v6_session}.

- bpf_map/publish_conntrack.rs: new submodule containing the two
  per-family helpers. Each is byte-identical to the original arm
  body — same struct field order, same byte-order conversions
  (.to_be() for ports, u32::from_ne_bytes for v4 NAT IPs, raw octets
  for v6 NAT IPs), same SESS_FLAG_* / SESS_STATE_ESTABLISHED
  constants, same early-return paths on cross-family reverse keys,
  same eprintln!("xpf-ha: conntrack vN map update failed: …")
  diagnostics on libbpf_sys failures.

No visibility bumps. The BPF struct types (BpfSessionKey/Value V4/V6)
and SESS_* constants stay private at bpf_map/mod.rs scope; the child
publish_conntrack module reaches them via Rust's parent-private-item
access rule.

No signature change on publish_bpf_conntrack_entry. All 5 call sites
(bpf_map.rs:839, 853 → now bpf_map/mod.rs:691, 705; poll_descriptor.rs:
915, 1482, 2782) continue to compile unchanged.

HA session-sync portability preserved: byte layout of BpfSessionValueV4
and BpfSessionValueV6 unchanged, .to_be() and u32::from_ne_bytes calls
in identical positions, BPF map FDs threaded through unchanged.

Validation:

- cargo build --release clean (114 pre-existing warnings, no new).
- cargo test --release: 1416 passed, 1 known wg flake
  (reconcile_peers_snapshot_is_atomic_under_concurrent_load) passes
  5/5 in isolation — unrelated to this change.
- Go suite: all 30 packages pass.

Per wave-2 batch policy (#1325/#1326/#1327/#1342/#1345/#1351/#1352)
no per-PR smoke; cluster smoke is batched at chain-end. Marker after
4-of-4 reviewer attestation (Codex + AGY + Copilot + Claude SMR).

Plan + adversarial review history at docs/pr/1356-bpf-map-split/.
Round-1: Codex PLAN-NEEDS-MAJOR (6 blockers, all about module/visibility
shape), AGY PLAN-NEEDS-MINOR (3 findings, overlap with Codex). Round-2
(v2 commit d179ff07): both PLAN-READY.

Closes #1356.
