+++ docs/pr/1356-bpf-map-split/reviewer-ids.md
@@ -6,6 +6,23 @@ here so the next iteration can fetch results by ID.
 
 ## Plan round 1 (commit 8398db39)
 
-- **Codex:** task-mpmypz53-viyk24
-- **AGY:** (pending — see /tmp/.../tasks/bhf22hdig.output)
+- **Codex:** task-mpmypz53-viyk24 — PLAN-NEEDS-MAJOR
+- **AGY:** review-mpmytock-qbgl6a — PLAN-NEEDS-MINOR
 - **Gemini:** not dispatched (4-of-4 attestation is Codex + AGY + Copilot + Claude SMR; no Gemini for this wave-2 batch)
+
+## Plan round 2 (commit d179ff07)
+
+- **Codex:** task-mpmz1e6v-iwubwq — PLAN-READY (provisional; sandbox could
+  not open files but reviewed design against v2 summary; remaining caveat
+  about explicit `super::` imports already addressed in plan §Imports)
+- **AGY:** review-mpmz1m34-0nf3ae — timed out exploring with cargo test
+- **AGY (retry):** review-mpmz5gkv-bof9lt — PLAN-READY
+
+Both reviewers PLAN-READY on v2.
+
+## Implementation (after PLAN-READY × 2)
+
+- Codex (code review): pending
+- AGY (code review): pending
+- Copilot: pending
+- Claude SMR: pending
